requirements.txt
beautifulsoup4
psycopg2
python-dotenv
requests